requirements.txt
accelerate==0.27.2
bitsandbytes==0.42.0
datasets==2.18.0
llama-index==0.10.16
llama-index-embeddings-huggingface==0.1.4
llama-index-embeddings-openai==0.1.6
llama-index-llms-huggingface==0.1.3
llama-index-llms-llama-cpp==0.1.3
llama-index-llms-openai==0.1.7
numpy==1.26.4
openai==1.13.3
pandas==2.2.1
pydantic==2.6.3
peft==0.9.0
scikit-learn
torch==2.1.2
tqdm==4.66.2
transformers==4.38.2
trl==0.7.11
uvicorn==0.27.1
fastapi==0.110.0
streamlit==1.32.1
streamlit-chat==0.1.1
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.32
langchain_openai==0.0.8
langchain-text-splitters==0.0.1
langchain_experimental==0.0.54
langchainhub==0.1.15
huggingface-hub==0.20.3
pypdf==4.1.0
openai==1.13.3
unstructured==0.12.6
unstructured[pdf]
tiktoken==0.6.0
faiss-cpu==1.8.0
chromadb==0.4.24
sentence-transformers==2.5.1
wikipedia==1.4.0
seaborn==0.13.2
langgraph==0.0.30
langserve
sse_starlette==2.0.0
pygraphviz==1.11
graphviz==0.20.3
# ubuntuÀÇ °æ¿ì 
# apt-get update
# apt-get install libgraphviz-dev
# pip install pygraphviz
pydot-ng==2.0.0
# sudo apt install nvidia-cuda-toolkit
# deepspeed
kor==1.0.1
sse_starlette==2.0.0